whatsapp_ingest/requirements.txt
aws-lambda-context==1.1.0
aws-lambda-logging==0.1.1
boto3==1.26.87
botocore==1.29.33
certifi==2022.12.7
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
emoji==2.2.0
mail-parser==3.15.0
python-dotenv==0.20.0
tika==1.24
vaderSentiment==3.3.2
